chore: move the tests into production code
