+++ src/core/CMakeLists.txt
@@ -25,6 +25,13 @@ target_link_libraries(
   KF5::ConfigGui
 )
 
+target_compile_definitions(
+  bismuth_core
+  PRIVATE
+  # Make all Doctest symbols exported
+  DOCTEST_CONFIG_IMPLEMENTATION_IN_DLL
+)
+
 kconfig_add_kcfg_files(bismuth_core GENERATE_MOC "config.kcfgc")
 
 install(
