+++ src/core/plasma-api/workspace.cpp
@@ -3,6 +3,10 @@
 
 #include "workspace.hpp"
 
+#include <doctest/doctest.h>
+
+#include "plasma-api/plasma-api.hpp"
+
 namespace PlasmaApi
 {
 
@@ -16,6 +20,27 @@ int Workspace::currentDesktop()
 {
     auto jsResult = m_jsRepr.property("currentDesktop");
     return jsResult.toInt();
+}
+
+namespace Test
+{
+
+TEST_CASE("Workspace Properties Read")
+{
+    auto engine = QQmlEngine();
+    auto mockWorkspace = MockWorkspaceJS();
+
+    engine.globalObject().setProperty(QStringLiteral("workspace"), engine.newQObject(&mockWorkspace));
+
+    auto plasmaApi = ::PlasmaApi::PlasmaApi(&engine);
+    auto workspace = plasmaApi.workspace();
+
+    SUBCASE("currentDesktop")
+    {
+        auto result = workspace.currentDesktop();
+        CHECK(result == 42);
+    }
+}
 }
 
 }
