--- tests/core/plasma-api/CMakeLists.txt
@@ -1,7 +1,0 @@
-# SPDX-FileCopyrightText: 2022 Mikhail Zolotukhin <mail@genda.life>
-# SPDX-License-Identifier: MIT
-
-target_sources(test_runner
-  PRIVATE
-  workspace.cpp
-)
